* misc/system.c (erealloc): Use z in format specifier for size_t, else 	compiler complains.
git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-libraries@20357 423fd1db-d629-0410-8442-d21db03e70f4

+++ misc/system.c
@@ -125,15 +125,14 @@ erealloc (void *ptr, size_t size)
 
   if ((int) size < 0)
     {
-      log_legacy_write ("Won't realloc() a pointer of size %ld !", size);
+      log_legacy_write ("Won't realloc() a pointer of size %zd !", size);
       exit (1);
     }
 
   ret = realloc (ptr, size);
   if (!ret)
     {
-      log_legacy_write ("Could not realloc() a pointer of size %ld !",
-                        (long) size);
+      log_legacy_write ("Could not realloc() a pointer of size %zd !", size);
       exit (1);
     }
   return ret;
